requirements.txt
streamlit~=1.36.0
numpy~=2.0.0
torch~=2.3.1
h5py~=3.11.0
opencv-python~=4.10.0.84
PyYAML~=6.0.1
hydra-core
omegaconf~=2.3.0
matplotlib~=3.9.1
packaging~=24.1
decorator~=5.1.1
gelsight~=4.0
setuptools~=68.2.0
pydantic~=2.8.2